% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/13_alphasimr_bridge.R
\name{alphasimr_make_hybrids}
\alias{alphasimr_make_hybrids}
\title{Cross the final elite lines to form the commercial hybrid pool}
\usage{
alphasimr_make_hybrids(elite_A, elite_B, SP)
}
\arguments{
\item{elite_A, elite_B}{Selected lines, see [alphasimr_rrs_cycle].}

\item{SP}{The `SimParam`.}
}
\value{
An AlphaSimR `Pop-class` of hybrids, `elite_A@nInd * elite_B@nInd`
  individuals.
}
\description{
A full factorial (testcross diallel) of the elite doubled-haploid lines
from each pool, phenotyped.
}
