% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomics.R
\name{qconfigMetabolomicsPlateXCalibur}
\alias{qconfigMetabolomicsPlateXCalibur}
\title{qconfigMetabolomics}
\usage{
qconfigMetabolomicsPlateXCalibur(x, howOften = 22, ...)
}
\arguments{
\item{df}{}
}
\value{
data.frame
}
\description{
qconfigMetabolomics
}
\section{Functions}{
\itemize{
\item \code{qconfigMetabolomicsPlateXCalibur()}: defined my MZ

}}
