% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pipeline_metrics.R
\name{coancestry_from_mrd}
\alias{coancestry_from_mrd}
\title{Molecular coancestry from modified Rogers distance}
\usage{
coancestry_from_mrd(D)
}
\arguments{
\item{D}{A modified Rogers distance matrix, see [mrd_matrix].}
}
\value{
The corresponding coancestry matrix, `1 - D^2`.
}
\description{
Molecular coancestry from modified Rogers distance
}
