% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_f2size.R
\name{cross_indices}
\alias{cross_indices}
\title{Generic cross between indexed individuals}
\usage{
cross_indices(pop, mo, fa, map, dist, max_try = 60)
}
\arguments{
\item{pop}{A population, a list of `H1` and `H2` haplotype matrices.}

\item{mo, fa}{Integer vectors of column indices. `mo == fa` is selfing.}

\item{map}{A map from [build_map].}

\item{dist}{Distortion specification from [make_distortion].}

\item{max_try}{Rejection attempts before accepting without zygotic selection.}
}
\value{
A new population.
}
\description{
Generic cross between indexed individuals
}
