% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pipeline_metrics.R
\name{dh_cross_coancestry}
\alias{dh_cross_coancestry}
\title{Coancestry between doubled haploids from two different families}
\usage{
dh_cross_coancestry(fL, i, j, k, l)
}
\arguments{
\item{fL}{Line coancestry matrix.}

\item{i, j}{Parent lines of the first cross.}

\item{k, l}{Parent lines of the second cross.}
}
\value{
A scalar.
}
\description{
`E[f] = (f_ik + f_il + f_jk + f_jl) / 4`.
}
\seealso{
Other DH identities:
\code{\link[=dh_pool_theta]{dh_pool_theta()}},
\code{\link[=dh_sib_coancestry]{dh_sib_coancestry()}},
\code{\link[=sigma2_dh]{sigma2_dh()}},
\code{\link[=sigma_dh_crosses]{sigma_dh_crosses()}}
}
\concept{DH identities}
