% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomics.R
\name{.autoQC01}
\alias{.autoQC01}
\title{autoQC01 template}
\usage{
.autoQC01(
  x,
  plateId = "1",
  QCrow = "H",
  mode = "",
  containerid = "",
  lc = "M_CLASS48_48"
)
}
\arguments{
\item{x}{}

\item{plateId}{}

\item{QCrow}{}

\item{mode}{}

\item{containerid}{}

\item{lssystem}{}
}
\value{
\code{data.frame} object
}
\description{
autoQC01 template
}
