% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomicsMclassVialXCaliburSII.R
\name{.autoQC01MclassVialXCaliburSII}
\alias{.autoQC01MclassVialXCaliburSII}
\title{autoQC01 Mclass Vial XCalibur SII}
\usage{
.autoQC01MclassVialXCaliburSII(
  x,
  plateId = "1",
  QCrow = "F",
  QCcol = 8,
  mode = "",
  containerid = "",
  lc = "Vanquish"
)
}
\arguments{
\item{x}{}

\item{plateId}{e.g., \code{"Y"}.}

\item{QCrow}{row plate coordinate, e.g., \code{"F"}.}

\item{QCcol}{column plate coordinate, e.g., \code{9}.}

\item{mode}{dia or dda}

\item{containerid}{}
}
\value{
\code{data.frame} object
}
\description{
autoQC01 Mclass Vial XCalibur SII
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2025-08-27
}
