% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomicsVialXCaliburSII.R
\name{.autoQC01VialXCaliburSII}
\alias{.autoQC01VialXCaliburSII}
\title{autoQC01 Vial XCaliburSII}
\usage{
.autoQC01VialXCaliburSII(
  x,
  plateId = "Y",
  QCrow = "F",
  QCcol = 9,
  mode = "",
  containerid = "",
  lc = "Vanquish"
)
}
\arguments{
\item{x}{}

\item{plateId}{e.g., \code{"Y"}.}

\item{QCrow}{row plate coordinate, e.g., \code{"F"}.}

\item{QCcol}{column plate coordinate, e.g., \code{9}.}

\item{mode}{dia or dda}

\item{containerid}{}

\item{lssystem}{}
}
\value{
\code{data.frame} object
}
\description{
autoQC01 Vial XCaliburSII
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2025-08-14
}
