% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomicsEquiSPLASH.R
\name{.blankEquiSPLASH}
\alias{.blankEquiSPLASH}
\title{blank Equi SPLASH}
\usage{
.blankEquiSPLASH(x, plateId = "Y", QCrow = "F", containerid = "", ...)
}
\arguments{
\item{x}{\code{data.frame} contains sample information, e.g., "Sample Name", "Sample ID", "Tube ID", "File Name", "Path", "Position"}

\item{plateId}{plate of the sample.}

\item{QCrow}{plate row of the sample.}

\item{containerid}{bfabric container, order or project id.}

\item{...}{}

\item{QCcol}{plate col of the sample.}
}
\value{
\code{data.frame}
}
\description{
blank Equi SPLASH
}
\examples{
take Inj Vol from x
}
\author{
Christian Panse <cp@fgcz.ethz.ch>, 2025-08-13
}
