% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.composePlateSampleTable}
\alias{.composePlateSampleTable}
\title{compose plate sample table}
\usage{
.composePlateSampleTable(
  p,
  orderID = 34843,
  area = "Metabolomics",
  mode = "",
  instrument = "ASTRAL_1",
  lc = "M_CLASS48_48",
  user = "cpanse",
  injVol = 3.5,
  plateCounter = 0,
  randomization = "plate"
)
}
\arguments{
\item{p}{}

\item{orderID}{}

\item{area}{}

\item{mode}{}

\item{instrument}{}

\item{user}{}

\item{injVol}{}

\item{plateCounter}{}

\item{randomization}{}
}
\description{
compose plate sample table
}
