% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.composeSampleTable}
\alias{.composeSampleTable}
\title{compose vial sample table}
\usage{
.composeSampleTable(
  x,
  orderID = 34843,
  area = "Metabolomics",
  mode = "",
  instrument = "ASTRAL_1",
  user = "cpanse",
  injVol = 3.5,
  randomization = TRUE
)
}
\description{
compose vial sample table
}
\examples{
.readSampleOfContainer(34843, login, webservicepassword, bfabricposturl) |> .composeSampleTable(orderID = 34843, randomization = TRUE) -> x
x|> qconfigMetabolomics()|> .replaceRunIds() -> xx
}
