% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.composeVialSampleTable}
\alias{.composeVialSampleTable}
\title{compose vial sample table}
\usage{
.composeVialSampleTable(
  x,
  orderID = 34843,
  area = "Metabolomics",
  mode = "",
  instrument = "ASTRAL_1",
  user = "cpanse",
  injVol = 3.5,
  lc = "M_CLASS48_48",
  randomization = TRUE
)
}
\arguments{
\item{x}{}

\item{orderID}{}

\item{area}{}

\item{mode}{}

\item{instrument}{}

\item{user}{}

\item{injVol}{}

\item{randomization}{}
}
\description{
compose vial sample table
}
\examples{
 if (all(c('login', 'webservicepassword', 'bfabricposturl') \%in\% names(Sys.getenv()))){
.readSampleOfContainer(36104,
    Sys.getenv('login'),
    Sys.getenv('webservicepassword'),
    Sys.getenv('bfabricposturl')) |>
  .composeVialSampleTable(orderID = 34843,
      lc = "M_CLASS48_48",
      randomization = FALSE) -> x
  x |> head()
  }
  
 if (all(c('login', 'webservicepassword', 'bfabricposturl') \%in\% names(Sys.getenv()))){
.readSampleOfContainer(34843,
    Sys.getenv('login'),
    Sys.getenv('webservicepassword'),
    Sys.getenv('bfabricposturl')) |>
  .composeVialSampleTable(orderID = 34843,
      randomization = TRUE,
      lc = "Vanquish") -> x
      x |> head()
      
# x|> qg:::qconfigMetabolomics()|> .replaceRunIds() -> xx
}
}
