% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomics.R
\name{.createMetabolomicsSample}
\alias{.createMetabolomicsSample}
\title{Generic builder to create metabolomics sample rows from config}
\usage{
.createMetabolomicsSample(x, sample_types)
}
\arguments{
\item{x}{input data frame with column structure}

\item{sample_types}{vector of sample type names to create}
}
\description{
Generic builder to create metabolomics sample rows from config
}
