% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.insertSample}
\alias{.insertSample}
\title{Generic function to place a sample into a queue configuration
\code{...} is used to pass the parameters to FUN}
\usage{
.insertSample(
  x,
  where = NA,
  howOften = round(nrow(x)/2),
  sampleFUN = NA,
  path = NA,
  ...
)
}
\arguments{
\item{x}{}

\item{where}{the position where the sample should be inserted}

\item{howOften}{how frequently the sample should be inserted}

\item{sampleFUN}{the function to generate the sample}

\item{path}{the path of the sample}

\item{...}{parameters to pass to sampleFUN}
}
\description{
Generic function to place a sample into a queue configuration
\code{...} is used to pass the parameters to FUN
}
