% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.interpolateFinalRows}
\alias{.interpolateFinalRows}
\title{Interpolates placeholders in filenames.}
\usage{
.interpolateFinalRows(x, container)
}
\arguments{
\item{x}{}
}
\description{
Interpolates placeholders in filenames.
}
