% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc-systems.R
\name{.lcVanquish}
\alias{.lcVanquish}
\title{Implementation of the Thermo Fisher Scientific Vanquish}
\usage{
.lcVanquish(n = 10)
}
\arguments{
\item{n}{number of vials}
}
\value{
a string vector of vial positions.
}
\description{
Implementation of the Thermo Fisher Scientific Vanquish
}
\examples{
.lcVanquish(96)
}
\references{
\url{https://www.thermofisher.com/ch/en/home/industrial/chromatography/liquid-chromatography-lc/hplc-uhplc-systems/vanquish-core-hplc-system.html}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2024-07-04
}
