% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc-systems.R
\name{.lcWater}
\alias{.lcWater}
\title{Implementation of the Water system `M_CLASS48x48`}
\usage{
.lcWater(n = 46)
}
\arguments{
\item{n}{number of vials}
}
\value{
a string vector of vial positions.
}
\description{
Implementation of the Water system `M_CLASS48x48`
}
\details{
works only for 46 vials (one plate); while expecting F7 and F8 are used for
clean and qc vials
}
\examples{
.lcWater(46)
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2024-09-04
}
