% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc-systems.R
\name{.lcWaters}
\alias{.lcWaters}
\title{Implementation of the Water system `M_CLASS48x48`}
\usage{
.lcWaters(n = 45)
}
\arguments{
\item{n}{number of vials}
}
\value{
a string vector of vial positions.
}
\description{
Implementation of the Water system `M_CLASS48x48`
}
\details{
works only for 45 vials (one plate); while expecting positions F6, F7 and F8
are used for clean, qc(autoQC03), qc(autoQC01) vials
}
\examples{
.lcWaters(45)
# two plates
.lcWaters(46)
testthat::expect_true(.lcWaters(47)[1] == "1:A,1")
testthat::expect_true(.lcWaters(47)[47] == "2:A,2")
testthat::expect_true(.lcWaters(47)[45] == "1:F,5")
testthat::expect_true(.lcWaters(90)[90] == "2:F,5")
testthat::expect_error(.lcWaters(91))
}
\references{
\url{https://www.waters.com/nextgen/ch/de/products/chromatography/chromatography-systems/acquity-uplc-m-class-system.html}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2024-09-04, Christian Panse <cp@fgcz.ethz.ch>, Claudia 2025-10-02
}
