% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomicsNew.R
\name{.metabolomicsQueueVial}
\alias{.metabolomicsQueueVial}
\title{Generic queue configuration for metabolomics vial-based workflows}
\usage{
.metabolomicsQueueVial(x, howOften, polarities, standard)
}
\arguments{
\item{x}{input data frame with primary samples and column structure}

\item{howOften}{integer, how often to insert QC samples}

\item{polarities}{character vector, polarities to include}

\item{standard}{character, standard to use}
}
\description{
Since at the queue level, the only difference between metabolomics and lipidomics
is the standard, setting the standard parameter is sufficient to cover both.
}
