% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomicsEquiSPLASH.R
\name{.pooledQCEquiSPLASH}
\alias{.pooledQCEquiSPLASH}
\title{take Inj Vol from x}
\usage{
.pooledQCEquiSPLASH(x, plateId = "Y", QCrow = "F", containerid = "", ...)
}
\description{
take Inj Vol from x
}
