% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.readSampleOfContainer}
\alias{.readSampleOfContainer}
\title{Read samples of a given container}
\usage{
.readSampleOfContainer(containerID, login, webservicepassword, posturl)
}
\arguments{
\item{containerID}{integer container ID}

\item{login}{character login}

\item{webservicepassword}{character webservice password you get from your bfabric user profile page}

\item{posturl}{character bfabric REST post url}
}
\description{
Read samples of a given container
}
\examples{
if (all(c('login', 'webservicepassword', 'bfabricposturl') \%in\% names(Sys.getenv()))){
  qg::.readSampleOfContainer(
    containerID = 35464,
    login = Sys.getenv('login'), 
    webservicepassword = Sys.getenv('webservicepassword'), 
    posturl = Sys.getenv('bfabricposturl')
    )
}
}
\author{
Christian Panse
}
