% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.toHystar}
\alias{.toHystar}
\title{generator Hystar config XML file}
\usage{
.toHystar(x, file = "file.xml")
}
\arguments{
\item{x}{a data frame}

\item{file}{output}
}
\value{
xml file
}
\description{
generator Hystar config XML file
}
\author{
Tobias Kockmann (intial), Claudia Fortes, Christian Panse
}
