% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomics.R
\name{.validatePlateNumber}
\alias{.validatePlateNumber}
\title{Vanquish plate number parser}
\usage{
.validatePlateNumber(x)
}
\arguments{
\item{x}{character position string}
}
\description{
Vanquish plate number parser
}
\details{
position string with a valid Vanquish plate numbering
}
\examples{
qg:::.parsePlateNumber("1:A4")
}
