% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc-systems.R
\name{.validateVanquishPlateNumber}
\alias{.validateVanquishPlateNumber}
\title{validateVanquishPlateNumber}
\usage{
.validateVanquishPlateNumber(x)
}
\description{
validateVanquishPlateNumber
}
\examples{
qg:::.parseVanquishPlateNumber("1:A4")
qg:::.lcVanquish() |> sapply(FUN = qg:::.parseVanquishPlateNumber)
}
