% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_mabc.R
\name{haldane_r}
\alias{haldane_r}
\title{Haldane recombination fraction}
\usage{
haldane_r(d_cM)
}
\arguments{
\item{d_cM}{Map distance in cM.}
}
\value{
The recombination fraction.
}
\description{
Haldane recombination fraction
}
