% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_f2size.R
\name{make_F1}
\alias{make_F1}
\title{Make an F1 population from two fully homozygous parents}
\usage{
make_F1(map, n)
}
\arguments{
\item{map}{A map from [build_map].}

\item{n}{Number of plants.}
}
\value{
A population.
}
\description{
Every F1 plant is genetically identical: `H1` is all P1 ancestry, `H2` all
P2. This is why the genetic `Ne` of the F1 is 1.
}
