% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_f2size.R
\name{marker_index}
\alias{marker_index}
\title{Index of the marker nearest a map position}
\usage{
marker_index(map, chr, pos_cM)
}
\arguments{
\item{map}{A map from [build_map].}

\item{chr}{Chromosome number.}

\item{pos_cM}{Position in cM.}
}
\value{
An integer marker index.
}
\description{
Index of the marker nearest a map position
}
