% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics.R
\name{mean_index}
\alias{mean_index}
\title{Mean multi-trait index of a selection}
\usage{
mean_index(idx, ctx)
}
\arguments{
\item{idx}{Integer vector of selected row indices.}

\item{ctx}{Evaluation context, see [build_ctx].}
}
\value{
A scalar, in standard deviations of the candidate pool.
}
\description{
Mean multi-trait index of a selection
}
