% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_fast_metrics.R
\name{ne_parents_from_counts}
\alias{ne_parents_from_counts}
\title{Effective number of parents from a count vector}
\usage{
ne_parents_from_counts(cnt)
}
\arguments{
\item{cnt}{Integer vector of line counts, see [line_counts].}
}
\value{
A scalar.
}
\description{
Effective number of parents from a count vector
}
\seealso{
Other fast metrics:
\code{\link[=alleles_retained_fast]{alleles_retained_fast()}},
\code{\link[=fast_ne_parents]{fast_ne_parents()}},
\code{\link[=fast_pbar]{fast_pbar()}},
\code{\link[=fast_theta]{fast_theta()}},
\code{\link[=fast_theta_pools]{fast_theta_pools()}},
\code{\link[=line_counts]{line_counts()}},
\code{\link[=line_usage_freq]{line_usage_freq()}},
\code{\link[=make_theta_fun]{make_theta_fun()}},
\code{\link[=pack_lines]{pack_lines()}},
\code{\link[=popcount32]{popcount32()}},
\code{\link[=theta_state]{theta_state()}},
\code{\link[=theta_swap]{theta_swap()}}
}
\concept{fast metrics}
