% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_metrics.R
\name{p_het}
\alias{p_het}
\title{Heterozygosity from an allele frequency}
\usage{
p_het(p)
}
\arguments{
\item{p}{Numeric vector of allele frequencies.}
}
\value{
`2 p (1 - p)`.
}
\description{
Heterozygosity from an allele frequency
}
