% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomics.R
\name{qconfigMetabolomicsPlateXCalibur}
\alias{qconfigMetabolomicsPlateXCalibur}
\title{qconfig metabolomics for plates}
\usage{
qconfigMetabolomicsPlateXCalibur(x, howOften = 22, ...)
}
\arguments{
\item{x}{a data.frame}

\item{howOften}{how frequently the sample should be inserted}

\item{...}{parameters to pass to \code{\link[qg]{.insertSample}}}
}
\value{
data.frame
}
\description{
qconfig metabolomics for plates
}
\details{
as defined my MZ
}
