% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomics.R
\name{qconfigMetabolomicsVanquishVialXCaliburSII_pos}
\alias{qconfigMetabolomicsVanquishVialXCaliburSII_pos}
\title{Queue configuration for metabolomics (positive mode only)}
\usage{
qconfigMetabolomicsVanquishVialXCaliburSII_pos(x, howOften)
}
\arguments{
\item{x}{input data frame with primary samples and column structure}

\item{howOften}{integer, how often to insert QC samples}
}
\value{
data frame with queue configuration
}
\description{
Queue configuration for metabolomics (positive mode only)
}
