% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configMetabolomicsEquiSPLASH.R
\name{qconfigMetabolomicsVanquishVialXCaliburSIIEquiSPLASH}
\alias{qconfigMetabolomicsVanquishVialXCaliburSIIEquiSPLASH}
\title{qconfig metabolomics for vials EquiSPLASH pos_neg}
\usage{
qconfigMetabolomicsVanquishVialXCaliburSIIEquiSPLASH(
  x,
  howOften = 8,
  mode = "",
  ...
)
}
\arguments{
\item{x}{a data.frame}

\item{howOften}{how frequently the sample should be inserted}

\item{...}{parameters to pass to \code{\link[qg]{.insertSample}}}
}
\description{
qconfig metabolomics for vials EquiSPLASH pos_neg
}
\details{
start with
- two blanks pos neg
- two EquiSPLASH
- two pooledQC
- two blanks
- two pooledQCDil
- two blanks
- samples pos neg
- ...
}
\examples{
qg:::.readPackageFile('test-Metabolomics-Vanquish-VialXCaliburSII-EXPLORIS_4-c37530.RData') -> S
S|> qconfigMetabolomicsVanquishVialXCaliburSIIEquiSPLASH(containerid = 37530) |> View()
}
\seealso{
also test-Metabolomics-Vanquish-VialXCaliburSII-EXPLORIS_4-c37530.R
}
