% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomicsMclassVialXCaliburSII.R
\name{qconfigProteomicsMclassVialXCaliburSIIdia}
\alias{qconfigProteomicsMclassVialXCaliburSIIdia}
\title{Proteomics MCLASS Vial XCalibur SII}
\usage{
qconfigProteomicsMclassVialXCaliburSIIdia(x, howOften = 8, ...)
}
\description{
Proteomics MCLASS Vial XCalibur SII
}
\author{
Christan Panse <cp@fgcz.ethz.ch> 2025-08-27
}
