% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomics.R
\name{qconfigProteomicsPlateChronos}
\alias{qconfigProteomicsPlateChronos}
\title{Proteomics Exploris EVOSEP 6x12x8 Plate Chronos}
\usage{
qconfigProteomicsPlateChronos(x, howOften = 4, ...)
}
\arguments{
\item{x}{a data.frame}

\item{howOften}{how frequently the sample should be inserted}

\item{...}{parameters to pass to \code{\link[qg]{.insertSample}}}
}
\value{
a data.frame object
}
\description{
Proteomics Exploris EVOSEP 6x12x8 Plate Chronos
}
\author{
Antje Dittmann 2024-09-03 2025-01-31
}
