% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomics.R
\name{qconfigProteomicsVialChronos}
\alias{qconfigProteomicsVialChronos}
\title{queue config for Proteomics Vial and Chronos}
\usage{
qconfigProteomicsVialChronos(x, howOften = 4, ...)
}
\arguments{
\item{x}{a data.frame}

\item{howOften}{how frequently the sample should be inserted}

\item{...}{parameters to pass to \code{\link[qg]{.insertSample}}}
}
\value{
data.frame
}
\description{
queue config for Proteomics Vial and Chronos
}
\details{
as defined my MZ
}
\author{
Martina CP 2024-11-05
}
