% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomicsVialXCaliburSII.R
\name{qconfigProteomicsVialXCaliburSII}
\alias{qconfigProteomicsVialXCaliburSII}
\title{Proteomics Vial XCalibur SII}
\usage{
qconfigProteomicsVialXCaliburSII(x, howOften = 8, ...)
}
\arguments{
\item{x}{}

\item{howOften}{integer, how often to insert QC samples}

\item{...}{}
}
\value{
data frame with queue configuration
}
\description{
Proteomics Vial XCalibur SII
}
\details{
similar setup as Astral - Vanquish - SII XCalibur

sample input see o39408
B:F9 autoQC01
B:F8 autoQC03
B:F7 clean
}
\examples{
S |> qconfigProteomicsVialXCaliburSII()
}
\author{
Antje Dittmann <antje.dittmann@fgcz.uzh.ch>,
Pfammatter Sibylle <sibylle.pfammatter@fgcz.ethz.ch>, 
Christan Panse <cp@fgcz.ethz.ch> 2025-08-13,14
}
