% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_pipeline_metrics.R
\name{sigma_dh_crosses}
\alias{sigma_dh_crosses}
\title{Progeny standard deviation for a set of candidate crosses}
\usage{
sigma_dh_crosses(crosses, GL, beta)
}
\arguments{
\item{crosses}{`C x 2` matrix of parent line indices.}

\item{GL}{Line genotype matrix coded 0/1.}

\item{beta}{Marker effects.}
}
\value{
A numeric vector of length `C`.
}
\description{
Vectorised form of [sigma2_dh]. Costs `O(C m)`.
}
\seealso{
Other DH identities:
\code{\link[=dh_cross_coancestry]{dh_cross_coancestry()}},
\code{\link[=dh_pool_theta]{dh_pool_theta()}},
\code{\link[=dh_sib_coancestry]{dh_sib_coancestry()}},
\code{\link[=sigma2_dh]{sigma2_dh()}}
}
\concept{DH identities}
