% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_mabc.R
\name{strat_rec_until}
\alias{strat_rec_until}
\title{Last generation in which recombinant selection is applied}
\usage{
strat_rec_until(strategy, opt)
}
\arguments{
\item{strategy}{One of the keys in `mabc_opt$strategies`.}

\item{opt}{Options, see [mabc_opt].}
}
\value{
An integer generation.
}
\description{
Last generation in which recombinant selection is applied
}
