% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_mabc.R
\name{target_indices}
\alias{target_indices}
\title{Marker indices of the first `k` targets}
\usage{
target_indices(map, k, targets = maize_targets)
}
\arguments{
\item{map}{A map from [build_map].}

\item{k}{Number of targets to use, from the top of `targets`.}

\item{targets}{A target table, see [maize_targets].}
}
\value{
An integer vector of `k` marker indices.
}
\description{
Marker indices of the first `k` targets
}
\examples{
map <- build_map(maize_chr_len, spacing_cM = 5)
target_indices(map, 2)
}
