% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_f2size.R
\name{zyg_fitness}
\alias{zyg_fitness}
\title{Zygotic fitness of proposed offspring}
\usage{
zyg_fitness(Gf, Gm, dist)
}
\arguments{
\item{Gf, Gm}{Female and male gamete matrices.}

\item{dist}{Distortion specification from [make_distortion].}
}
\value{
A numeric vector of relative fitnesses.
}
\description{
Zygotic fitness of proposed offspring
}
