% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.insertSample}
\alias{.insertSample}
\alias{Christian.Panse}
\alias{<cp@fgcz.ethz.ch>}
\alias{2023,}
\alias{2025-10-02}
\title{Generic function to place a sample into a queue configuration
\code{...} is used to pass the parameters to \code{sampleFUN} method.}
\usage{
.insertSample(
  x,
  modOffset = 0,
  where = NA,
  howOften = round(nrow(x)/2),
  sampleFUN = NA,
  path = NA,
  ...
)
}
\arguments{
\item{x}{\code{data.frame} contains sample information, e.g., \code{"Sample Name", "Sample ID", "Tube ID", "File Name", "Path", "Position"}.}

\item{where}{the position where the sample should be inserted}

\item{howOften}{how frequently the sample should be inserted}

\item{sampleFUN}{the implemented function used defining the sample, e.g., \link[qg]{.autoQC01VialXCaliburSII}.}

\item{path}{the path of the sample.}

\item{...}{parameters to pass to sampleFUN function.}
}
\description{
Generic function to place a sample into a queue configuration
\code{...} is used to pass the parameters to \code{sampleFUN} method.
}
