% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc-systems.R
\name{.lcWaters}
\alias{.lcWaters}
\title{Implementation of the Water system `M_CLASS48x48`}
\usage{
.lcWaters(n = 46)
}
\arguments{
\item{n}{number of vials}
}
\value{
a string vector of vial positions.
}
\description{
Implementation of the Water system `M_CLASS48x48`
}
\details{
works only for 46 vials (one plate); while expecting F7 and F8 are used for
clean and qc vials
}
\examples{
.lcWaters(46)
}
\references{
\url{https://www.waters.com/nextgen/ch/de/products/chromatography/chromatography-systems/acquity-uplc-m-class-system.html}
}
\author{
Christian Panse <cp@fgcz.ethz.ch> 2024-09-04
}
