% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configs.R
\name{.readSampleOfContainer}
\alias{.readSampleOfContainer}
\title{read samples of a given container}
\usage{
.readSampleOfContainer(containerID, login, webservicepassword, posturl)
}
\description{
read samples of a given container
}
\examples{
.readSampleOfContainer(35464, login, webservicepassword, bfabricposturl)
}
