% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configProteomics.R
\name{qconfigProteomicsEVOSEP6x12x8PlateHystar}
\alias{qconfigProteomicsEVOSEP6x12x8PlateHystar}
\title{queue confiug for Proteomics EVOSEP 6x12x8 Plate Hystar}
\usage{
qconfigProteomicsEVOSEP6x12x8PlateHystar(x, howOftenQC = 48, ...)
}
\description{
queue confiug for Proteomics EVOSEP 6x12x8 Plate Hystar
}
\details{
increments clean and qc positions
}
\author{
Claudia Fortes & Christian Panse
}
